Refactor layout and add new views for Account and Traffic
