+++ frontend/src/App.vue
@@ -4,38 +4,30 @@
         <el-header>
           <Menu />
         </el-header>
-        <el-main style="text-align: center;">
-          <Info />
-          <el-row style="margin-top: 20px" :gutter="20" :justify="'center'">
-            <el-col :span="4">
-              <setting />
-            </el-col>
-            <el-col :span="4">
-              <Monitor />
-            </el-col>
-            <el-col :span="4">
-              <Login />
-            </el-col>
-            <el-col :span="4">
-              <Logout />
-            </el-col>
-          </el-row>
+        <el-main style="text-align: center">
+          <el-tabs type="border-card" tab-position="top">
+            <el-tab-pane label="账号">
+              <Account/>
+            </el-tab-pane>
+            <el-tab-pane label="监控">
+              <Traffic/>
+            </el-tab-pane>
+            <el-tab-pane label="关于">
+              <About />
+            </el-tab-pane>
+          </el-tabs>
         </el-main>
         <el-footer style="text-align: center">
-          <About />
         </el-footer>
       </el-container>
   </div>
 </template>
 
 <script setup lang="ts">
-import Setting from "./components/Setting.vue";
-import Info from "./components/Info.vue";
-import Login from "./components/Login.vue";
-import Logout from "./components/Logout.vue";
 import Menu from "./components/Menu.vue";
-import About from "./components/About.vue";
-import Monitor from "./components/Monitor.vue";
+import About from "./views/About.vue";
+import Account from "./views/Account.vue";
+import Traffic from "./views/Traffic.vue";
 </script>
 
 <style scoped>
