+++ frontend/src/views/About.vue
@@ -3,11 +3,9 @@
 </script>
 
 <template>
-  <div class="footer">
-    <div class="about">
-      本项目采用 GPL-3.0 license 进行许可<br/>
-      仓库地址: https://github.com/HEX9CF/CupidConnector
-    </div>
+  <div class="about">
+    本项目采用 GPL-3.0 license 进行许可<br/>
+    仓库地址: https://github.com/HEX9CF/CupidConnector
   </div>
 </template>
