+++ frontend/src/views/Traffic.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <el-empty style="max-height: 320px"></el-empty>
+</template>
+
+<style scoped>
+
+</style>
